* remove old example widgets * refactor package structure of new widget
