+++ app/src/main/AndroidManifest.xml
@@ -57,10 +57,10 @@
         </activity>
 
         <service
-                android:name=".widget.thirdStab.TaskWidgetService"
+                android:name=".widget.TaskWidgetService"
                 android:permission="android.permission.BIND_REMOTEVIEWS"/>
 
-        <receiver android:name=".widget.thirdStab.TaskWidgetProvider">
+        <receiver android:name=".widget.TaskWidgetProvider">
             <intent-filter>
                 <action android:name=
                                 "android.appwidget.action.APPWIDGET_UPDATE"/>
@@ -68,7 +68,7 @@
 
             <meta-data
                     android:name="android.appwidget.provider"
-                    android:resource="@xml/example_appwidget_info"/>
+                    android:resource="@xml/appwidget_info"/>
         </receiver>
 
         <provider
