--- app/src/main/java/com/example/tasktrackerclient/widget/WidgetService.kt
@@ -1,16 +1,0 @@
-package com.example.tasktrackerclient.widget
-
-import android.content.Intent
-import android.widget.RemoteViewsService
-
-
-class WidgetService : RemoteViewsService() {
-
-    override fun onGetViewFactory(intent: Intent): RemoteViewsFactory {
-
-        return WidgetDataProvider(
-            applicationContext, intent
-        )
-    }
-
-}
