--- app/src/main/res/xml/new_app_widget_info.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-                    android:initialKeyguardLayout="@layout/new_app_widget"
-                    android:initialLayout="@layout/new_app_widget"
-                    android:minHeight="180dp"
-                    android:minWidth="110dp"
-                    android:previewImage="@drawable/example_appwidget_preview"
-                    android:resizeMode="horizontal|vertical"
-                    android:updatePeriodMillis="1800000"
-                    android:widgetCategory="home_screen">
-</appwidget-provider>
